# Optimize `array!` DSL (Round 2)